requirements.txt
Django==2.2
django-cors-headers==3.4.0
djangorestframework==3.11.1
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
